% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness_api.R
\name{compactness_api}
\alias{compactness_api}
\title{This function queries the compactness API rather than calculating compactness locally.}
\usage{
compactness_api(dir, namecol)
}
\arguments{
\item{dir}{The name of a directory holding the shp, prj, shx, and dbf files.}

\item{namecol}{The ID of the column with the district name in it}
}
\value{
A list of three: the metadata in n x p where n is the number of districts in the shapefile and p is the number of covariates; a list of lists of n where each sublist is a lat-long matrix, and a sublist of length >1 indicates a noncontiguous district; and the name column
}
\description{
This function queries the compactness API rather than calculating compactness locally.
}
\examples{
compactness_api(dir = "shapefiles", namecol = "GEOID")
}
