% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_predictions.R
\name{generate_predictions}
\alias{generate_predictions}
\title{Runs the main prediction, which is an ensemble as per Kaufman, King, and Komisarchik (2018)}
\usage{
generate_predictions(features, namecol)
}
\arguments{
\item{features}{A features object produced by the "generate_features" function}
}
\value{
A vector of length corresponding to number
  of polygons in the shp.
}
\description{
Runs the main prediction, which is an ensemble as per Kaufman, King, and Komisarchik (2018)
}
\examples{
filename = "CnclDist_July2012.shp"
namecol = "NAME"
shp = read_shapefiles(filename, namecol)
features = genreate_features(shp)
preds = generate_predictions(features, shp[[3]])
}
