% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness_wrapper.R
\name{get_compactness}
\alias{get_compactness}
\title{This is the main function in this package, and the only one you should use normally.}
\usage{
get_compactness(shp, namecol)
}
\arguments{
\item{shp}{The filename of a shp file containing district polygons.}

\item{namecol}{The name of the variable in the shapefile containing each district's unique identifier}
}
\value{
A vector of length corresponding to number
  of polygons in shp.
}
\description{
Estimates the compactness and two standard errors for each polygon in a shp file as per Kaufman, King, and Komisarchik (2018)
}
\examples{
get_compactness("CnclDist_July2012.shp")
}
